Changed FoldableBodyBag.cs into PlaceItemAsObject.cs. Placeable chairs.
fo real
