+++ UnityProject/Assets/Resources/Prefabs/BadVariants/Objects/Furniture/BVOld_wooden_chair.prefab
@@ -46,6 +46,7 @@ SpriteRenderer:
   m_LightProbeUsage: 0
   m_ReflectionProbeUsage: 0
   m_RayTracingMode: 0
+  m_RayTraceProcedural: 0
   m_RenderingLayerMask: 4294967295
   m_RendererPriority: 0
   m_Materials:
@@ -107,7 +108,7 @@ GameObject:
   - component: {fileID: -4125858894873895098}
   - component: {fileID: -1442809630732524517}
   m_Layer: 23
-  m_Name: wooden_chair
+  m_Name: BVOld_wooden_chair
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -173,6 +174,7 @@ MonoBehaviour:
   objectType: 1
   AtmosPassable: 1
   Passable: 1
+  passableExclusionsToThis: []
 --- !u!114 &114840956348795048
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -392,7 +394,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cc1f386f82014aa4ba117d2289c3d540, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  prefabVariant: {fileID: 1011433845357754, guid: dce8b95977b293b49974fcc26e4790c2,
+  prefabVariant: {fileID: 3175792500284053558, guid: 9843990c973070e40bb9244d59d676b1,
     type: 3}
 --- !u!114 &-4125858894873895098
 MonoBehaviour:
@@ -420,8 +422,6 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   syncMode: 0
   syncInterval: 0.1
-  isFlippable: 0
-  flippedObject: {fileID: 0}
 --- !u!1 &5276609779432803648
 GameObject:
   m_ObjectHideFlags: 0
@@ -468,6 +468,7 @@ SpriteRenderer:
   m_LightProbeUsage: 1
   m_ReflectionProbeUsage: 1
   m_RayTracingMode: 0
+  m_RayTraceProcedural: 0
   m_RenderingLayerMask: 1
   m_RendererPriority: 0
   m_Materials:
@@ -490,7 +491,7 @@ SpriteRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: -932335445
-  m_SortingLayer: 17
+  m_SortingLayer: 18
   m_SortingOrder: 0
   m_Sprite: {fileID: 0}
   m_Color: {r: 1, g: 1, b: 1, a: 1}
