+++ UnityProject/Assets/Resources/Prefabs/Objects/Furniture/Chairs/BarStoolItem.prefab
@@ -30,6 +30,12 @@ PrefabInstance:
       value: 
       objectReference: {fileID: 11400000, guid: 98136b65c48e5a94b85a680617fb288d,
         type: 2}
+    - target: {fileID: 7688367562615853652, guid: 5615206684c0a714997f9e856ba438ce,
+        type: 3}
+      propertyPath: prefabVariant
+      value: 
+      objectReference: {fileID: 8667645117007165039, guid: 9203bf917337c3e4db0a64a10ef0e50b,
+        type: 3}
     - target: {fileID: 7803598227193638417, guid: 5615206684c0a714997f9e856ba438ce,
         type: 3}
       propertyPath: m_Sprite
