+++ UnityProject/Assets/Resources/Prefabs/Objects/Furniture/Chairs/StoolItem.prefab
@@ -30,6 +30,12 @@ PrefabInstance:
       propertyPath: initialName
       value: stool
       objectReference: {fileID: 0}
+    - target: {fileID: 6573704770868306816, guid: c8af057edecfa6c4f877ae1b042d6ead,
+        type: 3}
+      propertyPath: prefabVariant
+      value: 
+      objectReference: {fileID: 8250781856238894824, guid: 177ff6e5fbec9d84bbda485c726757e8,
+        type: 3}
     - target: {fileID: 6756277979569020869, guid: c8af057edecfa6c4f877ae1b042d6ead,
         type: 3}
       propertyPath: m_Sprite
