+++ UnityProject/Assets/Resources/Prefabs/Objects/Furniture/Chairs/WoodenChairItem.prefab
@@ -30,6 +30,12 @@ PrefabInstance:
       value: 
       objectReference: {fileID: 11400000, guid: e8e65e72e99582e4ab0e0d00921bcee1,
         type: 2}
+    - target: {fileID: 6573704770868306816, guid: c8af057edecfa6c4f877ae1b042d6ead,
+        type: 3}
+      propertyPath: prefabVariant
+      value: 
+      objectReference: {fileID: 5346880051612449098, guid: ce971a77eb7c2894e8583efec1c56965,
+        type: 3}
     - target: {fileID: 6756277979569020869, guid: c8af057edecfa6c4f877ae1b042d6ead,
         type: 3}
       propertyPath: m_Sprite
