+++ UnityProject/Assets/Resources/Prefabs/Objects/Furniture/Chairs/WoodenChairWingedItem.prefab
@@ -7,6 +7,12 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
+    - target: {fileID: 2889685064118359855, guid: 9843990c973070e40bb9244d59d676b1,
+        type: 3}
+      propertyPath: prefabVariant
+      value: 
+      objectReference: {fileID: 2465317846002368811, guid: 65778087a63ccbd428d4033cbeca6663,
+        type: 3}
     - target: {fileID: 3170864351630566722, guid: 9843990c973070e40bb9244d59d676b1,
         type: 3}
       propertyPath: m_LocalPosition.x
